Web/requirements.txt
streamlit
pandas
numpy
streamlit_lottie
requests
streamlit-extras
streamlit-annotated-text
Pillow
matplotlib
plotly
